+++ benches/criterion/benchmarks/base.rs
@@ -23,8 +23,7 @@ fn push_pop_local(b: &mut Bencher) {
 
     b.iter(|| {
         prod.push(1).unwrap();
-        black_box(cons.peek_ref().unwrap());
-        unsafe { cons.advance(1); }
+        black_box(cons.pop().unwrap());
     });
 }
 
@@ -36,8 +35,7 @@ fn push_pop_shared(b: &mut Bencher) {
 
     b.iter(|| {
         prod.push(1).unwrap();
-        black_box(cons.peek_ref().unwrap());
-        unsafe { cons.advance(1); }
+        black_box(cons.pop().unwrap());
     });
 }
 
@@ -53,8 +51,7 @@ fn push_pop_x100(b: &mut Bencher) {
             prod.push(1).unwrap();
         }
         for _ in 0..BATCH_SIZE {
-            black_box(cons.peek_ref().unwrap());
-            unsafe { cons.advance(1); }
+            black_box(cons.pop().unwrap());
         }
     });
 }
@@ -71,8 +68,7 @@ fn push_pop_x100_local(b: &mut Bencher) {
             prod.push(1).unwrap();
         }
         for _ in 0..BATCH_SIZE {
-            black_box(cons.peek_ref().unwrap());
-            unsafe { cons.advance(1); }
+            black_box(cons.pop().unwrap());
         }
     });
 }
@@ -102,8 +98,7 @@ fn push_pop_work(b: &mut Bencher) {
             }
         }
         for _ in 0..BATCH_SIZE {
-            black_box(cons.peek_ref().unwrap());
-            unsafe { cons.advance(1); }
+            black_box(cons.pop().unwrap());
         }
     });
 }